chore: setAssetsToDb moved and renamed
